feat: /counseling-complete ページを追加 (#40)
/counseling-complete-lp-1 と同一内容のサンクスページをオーナー指示で追加。
本体を共通コンポーネント CounselingCompleteSection に抽出して両ページで共用し、
OpenAI Ads の CV 計測も両ページで発火する(旧サイトの挙動を踏襲)。
noindex・sitemap 除外。仕様書を更新。
