+++ src/app/sitemap.ts
@@ -6,7 +6,7 @@ import { getCareerPathData } from '@/lib/getCareerPathData';
 import { getTotalPages } from '@/utils/pagination';
 
 export default function sitemap(): MetadataRoute.Sitemap {
-  // noindex ページは除外: サンクスページ(counseling/complete, counseling-complete-lp-1)・広告LP(lp-1)
+  // noindex ページは除外: サンクスページ(counseling/complete, counseling-complete, counseling-complete-lp-1)・広告LP(lp-1)
   const staticPaths = [
     pages.index.url,
     pages.courses.url,
